requirements.txt
alembic==1.16.5
amqp==5.3.1
anyio==4.10.0
APScheduler==3.11.1
billiard==4.2.2
boto3==1.34.0
botocore==1.34.162
CacheControl==0.14.3
cachetools==5.5.2
celery==5.5.3
certifi==2022.9.24
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.6
Faker==37.12.0
firebase_admin==7.1.0
Flask==2.2.2
flask-cors==6.0.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.2
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.74.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.1.1
proto-plus==1.26.1
protobuf==6.31.1
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
redis==6.4.0
requests==2.28.1
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.44
tabulate==0.9.0
typing_extensions==4.14.1
urllib3==1.26.12
uvicorn==0.38.0
Werkzeug==2.2.2